requirements.txt
apache-airflow[dev]==2.6.0